api/requirements.txt
amqp==5.1.1
asgiref==3.5.2
astroid==2.12.13
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.0
billiard==3.6.4.0
boto3==1.26.94
boto3-stubs==1.26.94
botocore==1.29.94
botocore-stubs==1.29.94
celery==5.2.7
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coverage==7.0.5
dill==0.3.6
Django==4.1.3
django-celery-beat==2.4.0
django-celery-results==2.4.0
django-filter==22.1
django-timezone-field==5.0
djangorestframework==3.14.0
exceptiongroup==1.1.0
factory-boy==3.2.1
Faker==16.6.0
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
isort==5.10.1
jmespath==1.0.1
kombu==5.2.4
lazy-object-proxy==1.8.0
mccabe==0.7.0
mypy-boto3-cloudformation==1.26.60
mypy-boto3-dynamodb==1.26.87
mypy-boto3-ec2==1.26.91
mypy-boto3-lambda==1.26.80
mypy-boto3-rds==1.26.72
mypy-boto3-s3==1.26.62
mypy-boto3-sqs==1.26.0.post1
numpy==1.24.2
packaging==23.0
platformdirs==2.5.4
pluggy==1.0.0
prompt-toolkit==3.0.36
pycodestyle==2.10.0
pylint==2.15.8
pylint-django==2.5.3
pylint-plugin-utils==0.7
pytest==7.2.1
pytest-cov==4.0.0
pytest-django==4.5.2
pytest-factoryboy==2.5.1
python-crontab==2.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
redis==4.4.2
requests==2.31.0
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tomlkit==0.11.6
types-awscrt==0.16.12
types-s3transfer==0.6.0.post6
typing_extensions==4.4.0
tzdata==2022.7
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
wrapt==1.14.1
